Delete forms for old wordSections. Move other lexicon forms into a forms/ folder
