+++ .vscode/settings.json
@@ -23,6 +23,7 @@
   "cSpell.words": [
     "conlang",
     "Conlang",
+    "conlangers",
     "conlanging",
     "conlangs",
     "Conlangs",
